This is VirtualC64 6.1

+++ VCCore/rvlib/Devices/LinearDevice.cpp
@@ -21,7 +21,7 @@ using utl::Range;
 std::vector<std::string>
 LinearDevice::describeDevice() const noexcept
 {
-    double b  = size();
+    double b  = double(size());
     double kb = b  / 1024.0;
     double mb = kb / 1024.0;
     double gb = mb / 1024.0;
