+++ VCCore/rvlib/FileSystems/Amiga/FileSystem.cpp
@@ -79,7 +79,7 @@ FileSystem::dumpState(std::ostream &os) const noexcept
 
     if (isFormatted()) {
 
-        auto fill = 100.0 * st.usedBlocks / st.traits.blocks;
+        auto fill = 100.0 * double(st.usedBlocks) / double(st.traits.blocks);
 
         os << std::setw(5) << std::left << ("DOS" + std::to_string(isize(traits.dos)));
         os << "  ";
@@ -116,7 +116,7 @@ FileSystem::dumpProps(std::ostream &os) const noexcept
 
     auto st   = stat();
     auto bst  = bootStat();
-    auto fill = 100.0 * st.usedBlocks / st.traits.blocks;
+    auto fill = 100.0 * double(st.usedBlocks) / double(st.traits.blocks);
 
     os << tab("Name");
     os << st.name << std::endl;
